Initial migration to/with ModernUI 3.11
There are bugs and many that do not work well

+++ common/src/main/java/icyllis/modernui/mc/text/StandardFontSet.java
@@ -183,7 +183,7 @@ public class StandardFontSet extends FontSet {
                     // bake glyph ourselves
                     var glyph = TextLayoutEngine.getInstance().lookupGlyph(
                             outlineFont,
-                            mStandardPaint.getFontSize(),
+                            (int) mStandardPaint.getFontSize(),
                             glyphs.getInt(0)
                     );
                     if (glyph != null) {
